Simplify handling of null client in pubsub.Topic.Subscribe().
